DEV: Move DeadSignal to conftest
